backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
SQLAlchemy==2.0.25
alembic==1.13.1
python-multipart==0.0.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1
face-recognition==1.3.0
numpy==1.26.4
Pillow==10.2.0
requests==2.31.0
pydantic-settings==2.2.1
email-validator==2.1.0.post1

